+++ WinFormsApp1/Donador.cs
@@ -18,6 +18,7 @@ namespace WinFormsApp1
             InitializeComponent();
             conn.Open();
             dataGridDonador.DataSource = Consultar();
+            conn.Close();
         }
         public DataTable Consultar()
         {
@@ -41,14 +42,38 @@ namespace WinFormsApp1
         }
         public void Eliminar(string n)
         {
-            string query = "Delete from \"donador\"" + "where \"curp\"='" + n + "'";
-            NpgsqlCommand ejecutor = new NpgsqlCommand(query, conn);
-            ejecutor.ExecuteNonQuery();
+            NpgsqlCommand cm = new NpgsqlCommand("Select curp from donador where curp = '" + n + "'", conn);
+            conn.Open();
+            NpgsqlDataReader dr = cm.ExecuteReader();
+            if (dr.Read())
+            {
+                conn.Close();
+                conn.Open();
+                string query = "Delete from \"donador\"" + "where \"curp\"='" + n + "'";
+                NpgsqlCommand ejecutor = new NpgsqlCommand(query, conn);
+                ejecutor.ExecuteNonQuery();
+                MessageBox.Show("Registro Eliminado!");
+                this.Close();
+            }
+            else
+            {
+                MessageBox.Show("Usuario no existe!");
+                conn.Close();
+
+            }
 
         }
         private void button3_Click(object sender, EventArgs e)
         {
-            dataGridDonador.DataSource = Consultar_Nombre(txt_Consulta_Don.Text);
+            if (string.IsNullOrEmpty(txt_Consulta_Don.Text))
+            {
+                MessageBox.Show("Favor de ingresar la CURP!");
+            }
+            else
+            {
+                dataGridDonador.DataSource = Consultar_Nombre(txt_Consulta_Don.Text);
+            }
+            
         }
 
         private void btn_modificar_Click(object sender, EventArgs e)
@@ -59,10 +84,21 @@ namespace WinFormsApp1
             }
             else
             {
-                Agregar_donador ag = new Agregar_donador();
-                ag.bandera(true);
-                ag.Show();
-                this.Close();
+                NpgsqlConnection conn = new NpgsqlConnection("Server= localhost; database=bloodDonation+; User Id= postgres; Password=danielleon");
+                NpgsqlCommand cm = new NpgsqlCommand("Select curp from donador where curp = '" + txt_Consulta_Don.Text + "'", conn);
+                conn.Open();
+                NpgsqlDataReader dr = cm.ExecuteReader();
+                if (dr.Read())
+                {
+                    Agregar_donador ag = new Agregar_donador();
+                    ag.bandera(true);
+                    ag.Show();
+                    this.Close();
+                }
+                else
+                {
+                    MessageBox.Show("Usuario no existe!");
+                }
             }
         }
 
@@ -75,8 +111,7 @@ namespace WinFormsApp1
             else
             {
                 Eliminar(txt_Consulta_Don.Text);
-                MessageBox.Show("Registro Eliminado!");
-                this.Close();
+                
             }
         }
     }
